backend/requirements.txt
# Core Django
Django==5.0.14
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Development
python-decouple==3.8

# Monitoring and Error Tracking
sentry-sdk[django]==2.37.0
django-silk==5.2.0

# Login protection (feature-flagged)
django-axes==6.5.0

# For naming validation
dataclasses==0.6

# Excel import/export
openpyxl==3.1.2

# Caching (Redis)
redis==5.0.1

# Production WSGI server
gunicorn==21.2.0
gevent==23.9.1

# Background jobs
celery[redis]==5.4.0

# OpenAPI schema generation
drf-spectacular==0.27.2
